login-frontend/requirements.txt
aiofiles
fastapi
jinja2
numpy
pandas
uvicorn
psutil
python-dateutil
python-dotenv